moved JavaNioUtils #102
